optimization: 优化 gcloud.taskflow3 目录结构
