+++ gcloud/taskflow3/apis/django/v4/__init__.py
@@ -10,21 +10,3 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-
-import sys
-from django.apps import AppConfig
-
-
-class Taskflow3Config(AppConfig):
-    name = "gcloud.taskflow3"
-    verbose_name = "GcloudTaskflow3"
-
-    def ready(self):
-        from gcloud.taskflow3.signals.handlers import pipeline_finish_handler  # noqa
-        from gcloud.taskflow3.signals.handlers import pipeline_revoke_handler  # noqa
-        from gcloud.taskflow3.signals.handlers import pipeline_fail_handler  # noqa
-
-        # bamboo-pipeline 流程实例 unpickle 路径兼容
-        from gcloud.taskflow3.domains import context
-
-        sys.modules["gcloud.taskflow3.context"] = context
