+++ gcloud/taskflow3/apis/django/v4/node_action.py
@@ -19,7 +19,7 @@ from django.views.decorators.http import require_POST
 from gcloud.utils.decorators import request_validate
 from gcloud.iam_auth.intercept import iam_intercept
 from gcloud.taskflow3.models import TaskFlowInstance
-from gcloud.taskflow3.validators import NodeActionV2Validator
+from gcloud.taskflow3.apis.django.validators import NodeActionV2Validator
 from gcloud.iam_auth.view_interceptors.taskflow import NodeActionV2Inpterceptor
 from gcloud.contrib.operate_record.decorators import record_operation
 from gcloud.contrib.operate_record.constants import RecordType, OperateType
