+++ gcloud/tests/taskflow3/queues/test_prepare_and_start_task_queue_resolver.py
@@ -14,7 +14,7 @@ specific language governing permissions and limitations under the License.
 from django.test import TestCase
 
 from gcloud.tests.mock_settings import *  # noqa
-from gcloud.taskflow3.queues import PrepareAndStartTaskQueueResolver
+from gcloud.taskflow3.domains.queues import PrepareAndStartTaskQueueResolver
 
 
 class PrepareAndStartTaskQueueResolverTestCase(TestCase):
